+++ resources/migrations/20191010114000-drop-start-end-from-form-template-resource-workflow.down.sql
@@ -10,6 +10,7 @@ ALTER TABLE form_template
 ADD COLUMN start timestamp with time zone NOT NULL DEFAULT CURRENT_TIMESTAMP,
 ADD COLUMN endt timestamp with time zone NULL DEFAULT NULL;
 --;;
-DROP INDEX resource_resid_u;
+DROP INDEX IF EXISTS resource_resid_u;
 --;;
-CREATE UNIQUE INDEX resource_resid_u ON resource (organization, resid, coalesce(endt, '10000-01-01'));
+-- We used to create a better resource_resid_u constraint here, but don't anymore.
+-- See rationale in 20200204064800-drop-unique-resid.down.sql